refactor(ffi): Plan 212 Phase 4 — technical debt cleanup
- Remove serde_json hardcoded special-case, replace with returns_result
  flag + well-known body_override registry for generic Result handling
- Add returns_result field to FunctionShim for syn-scanned Result types
- Rename auto_cache.rs → module_cache.rs to avoid shadowing external crate
- Pass dep features to Cargo.toml generation (features field was parsed
  but never used)
- Add Sandbox::gc() for cleaning old cached crates and build artifacts
- Fix Math.min/max/abs i64→i32 to match Auto int type
- Clean dead code in ffi.rs (register_rust_function, create_rust_shim)
- Optimize manifest JSON embedding with raw strings

+++ crates/auto-lang/src/vm/ffi/stdlib.rs
@@ -1126,17 +1126,17 @@ pub fn shim_option_or(task: &mut AutoTask, vm: &AutoVM) -> Result<(), VMError> {
 // ============================================================================
 
 #[auto_macros::rust_fn("Math.abs", "auto.math.abs")]
-pub fn shim_math_abs(n: i64) -> i64 {
+pub fn shim_math_abs(n: i32) -> i32 {
     n.abs()
 }
 
 #[auto_macros::rust_fn("Math.min", "auto.math.min")]
-pub fn shim_math_min(a: i64, b: i64) -> i64 {
+pub fn shim_math_min(a: i32, b: i32) -> i32 {
     a.min(b)
 }
 
 #[auto_macros::rust_fn("Math.max", "auto.math.max")]
-pub fn shim_math_max(a: i64, b: i64) -> i64 {
+pub fn shim_math_max(a: i32, b: i32) -> i32 {
     a.max(b)
 }
